KIT: Implemented Automatic Performance Balancing to make a execute continuous event take about 1s for an infinite loop on no matter which platform.

+++ src/commonMain/kotlin/emulator/kit/Architecture.kt
@@ -101,7 +101,7 @@ abstract class Architecture(config: Config, asmConfig: AsmConfig) {
      * Execution Event: multi step
      * should be implemented by specific archs
      */
-    open fun exeMultiStep(steps: Int) {
+    open fun exeMultiStep(steps: Long) {
         getConsole().clear()
     }
